fix: components folder structure
