+++ client/src/components/Article/Actions/EditArticle.tsx
@@ -10,7 +10,7 @@ import {
   CardFooter,
   Button,
 } from "@chakra-ui/react";
-import { blue_color, white_color } from "../../assets/customColors";
+import { blue_color, white_color } from "../../../assets/customColors";
 
 export default function EditArticle() {
   const [title, setTitle] = useState("This is a dummy title");
